chore: saving non-scalar function impls for future work
In addition to get/set/delete, a large number of operations for collection data types are mostly implemented. This branch contains the implementations and tests for the additional methods. They need to be double checked for correctness and updated to include type annotations for mypy. Tests for each method are limited to happy path testing and should be expanded upon.