fix: don't run README generator in test matrix
Prior to this commit, the README generator was being run
inside of the test step, which gets run on several different
versions of python. Running the README generator more than
once causes issues because the upstream branch ends up with
divergent commits in it.

This commit pulls the README generator out to be its own
job in the workflow.